Initial implementation of late lambda-lifting
At this time, it's largely a proof of concept, since the results aren't good.
The decision procedure is about as simple as it gets: If a binding already takes
at least one value parameter, and afterward it will take at most ten value
parameters, we lift. This only scratches the surface of the heuristics done by
the wip/llf branch, but it's a start.