Abstract tyvars free in type but not in command
If a command ends in an empty Case, its type may not appear at all in its
syntax. Since MFE floating is written in terms of commands, the variables to
abstract are computed by lvlMFEComm, which takes the return type as an argument;
thus it needs to add the tyvars free in that type to the free vars of the
command.