Don't float MFE commands with free join points
This sounds obvious now that I say it out loud, so to speak. Anyway, it never
made that much sense to talk about floating commands as if they're terms.
Really, we should be floating MFE terms as terms and MFE commands as join
points. Not being able to float join points more often sometimes interacts badly
with the rather coarse join points currently created in translation; pitting our
Float Out versus GHC's, we lose by 7% in allocations in the
spectral/hartel/sched test because something ends up lazy instead of strict.

We could also be losing there because we do case-of-case early (because it's
easier to *do* it than *not* to in Sequent Core); it's not clear from inspection
where the problem begins.