Translate: Correct handling of case-of-let-join
The way fromCoreExpr piles up bindings is equivalent to translating more naively
and then eliminating a bunch of administrative redexes (in the sequent calculus,
these are exactly <compute ... | ...> forms). In particular, case-of-let looks
like:

    <compute (let x = v in c) | case ...>
      =>
    let x = v in <compute c | case ...>

(If we stop after lifting x.) This is valid because the continuation is not
bound in v, because v is a term. But if x were a join variable, this would *not*
be okay because the continuation variable may be free in x's RHS and hence now
bound to some outer continuation.

So when we hit a let that we're translating as a join point, we have to proceed
much as for a case, deciding whether the continuation is trivial and copying it
or constructing a redex (so the simplifier can deal with it).