Optionally float nullary join points
A nullary join point will become a thunk, potentially increasing sharing at the
cost of allocation. This is the same calculus behind full laziness, so it seems
worth at least trying (despite a comment in GHC's Simplify to the effect that
floating nullary joins never helps). It doesn't have any effect in any of the
spectral tests, anyway.