Zap demand info at top level
It's not clear why this never has to be done explicitly by GHC FloatOut, since
in principal it's perfectly reasonable that a strict let might float to top
level. It seems something has already zapped the demand from strict lets by the
time FloatOut gets to it. At any rate, for LLL we've let-bound all nontrivial
arguments and annotated them with the strictness from the function, so strict
lets can and do crop up.