Float scrutinees from case statements
In the rather messy way that I reorganized SetLevels to work with Sequent Core,
I neglected to float the scrutinees from case statements (which is a much less
natural thing to do in Sequent Core than in Core). This seemed to have a
particularly nasty effect when it caused CSE to fail (since only things floated
to the same level get CSE'd).