Simplifier: Don't eat dead continuations if trivial
When a function is known to be bottoming, anything that happens after it is
invoked is dead code. Hence (like stock GHC) we chuck any continuation,
including further arguments and case analysis, following the saturation of a
bottoming function. However, this doesn't really help if you're only throwing
away casts and type arguments, and it appears that strictness analysis can get
screwed up in that case, with small but measurable effects (+1.8% allocs in
wave4main).