WIP
Finally got aggressive contification working satisfactorily. Simplified things a
bit by adding the concepts of a continuation's disposition and a call's region;
the CallRegion type improved internal APIs, and the Disposition concept allows
us to say that a continuation shouldn't be floated into but *also* shouldn't be
counted as escaping (used for unfoldings and rules).