Bring ExprSize in line with baseline GHC
To isolate changes due to the new architecture, we want the new code to mimic
the old code as quickly as possible. These changes make sure that the unfolding
guidance we calculate is always the same one that baseline GHC would. Also I
added a warning whenever the calculation differs.