Don't call ensureDupableKont unless necessary
Ideally, we would know when it's necessary because the occurrence analyzer would
count occurrences of Return, but currently it has nowhere to record it because
Compute doesn't carry a binder with OccInfo the way term binders do. The new
commandDuplicatesKont and endDuplicatesKont functions do an ad hoc analysis, but
it's potentially slow.